Finished disassembly of server.js into db/sql.js.
